project/backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
openai>=1.12
python-dotenv>=1.0
